requirements.txt
mysql-connector-python
streamlit
pandas
numpy
weasyprint
datetime
requests
ms-graph-python-client